requirements.txt
bleak
paho-mqtt
flask
numpy
pillow
